fix(forge): Regenerate 16 C++ codec goldens for narrowing-cast wrap
- Wrap multi-byte decode in static_cast<carrier_t>(...) per generator.rs (15 codec goldens)
- Drop elide_lower checks on uint8/uint16 vs zero/T_MAX (crossfile_validator_codec golden)
- sce-build cargo test: 1639 PASS / 16 FAIL -> 1672 PASS / 0 FAIL

+++ tests/forge/expected/codec_init_syn_body.h
@@ -59,7 +59,7 @@ struct CodecInitSynBody {
         if ((parent_flags & 0x40) != 0) {
             const std::uint8_t* raw = cursor.peek_slice(2);
             if (raw == nullptr) return std::nullopt;
-            batch_size = static_cast<uint16_t>((static_cast<uint16_t>(raw[0]) << 8) | raw[1]);
+            batch_size = static_cast<uint16_t>(static_cast<uint16_t>((static_cast<uint16_t>(raw[0]) << 8) | raw[1]));
             if (!cursor.advance(2)) return std::nullopt;
         }
         return CodecInitSynBody{
