+++ tests/forge/expected/codec_repeat_elem.h
@@ -28,7 +28,7 @@ struct CodecRepeatElem {
         const std::uint8_t* raw = cursor.peek_slice(2);
         if (raw == nullptr) return std::nullopt;
         CodecRepeatElem value{
-            .seq = (static_cast<uint16_t>(raw[0]) << 8) | raw[1],
+            .seq = static_cast<uint16_t>((static_cast<uint16_t>(raw[0]) << 8) | raw[1]),
         };
         if (!cursor.advance(2)) return std::nullopt;
         return value;
