+++ tests/forge/expected/codec_zenoh_join.h
@@ -79,7 +79,7 @@ struct CodecZenohJoin {
         if ((parent_flags & 0x40) != 0) {
             const std::uint8_t* raw = cursor.peek_slice(2);
             if (raw == nullptr) return std::nullopt;
-            batch_size = static_cast<uint16_t>((static_cast<uint16_t>(raw[0]) << 8) | raw[1]);
+            batch_size = static_cast<uint16_t>(static_cast<uint16_t>((static_cast<uint16_t>(raw[0]) << 8) | raw[1]));
             if (!cursor.advance(2)) return std::nullopt;
         }
         auto lease_opt = cursor.read_vle_u64();
